Integrated with Rx

+++ src/Tick/ViewResponse.cs
@@ -3,8 +3,8 @@ namespace Tick.Response
 	public class ViewResponse
 	{	
 		public string Id{get;set;}
-		public int Amount{get;set;}
-		public int Price{get;set;}		
-		public int MarketValue{get;set;}
+		public decimal Amount{get;set;}
+		public decimal Price{get;set;}		
+		public decimal MarketValue{get;set;}
 	}
 }
